feat: file structure according to white tiger guideline, minor fixes
