+++ white_tiger_shop/lib/cart/model/cart_model.dart
@@ -1,6 +1,6 @@
 import 'package:hive/hive.dart';
 import 'package:hive_flutter/adapters.dart';
-import 'package:white_tiger_shop/common/model/base_model.dart';
+import 'package:white_tiger_shop/core/model/base_model.dart';
 import 'package:white_tiger_shop/product/model/entity/product.dart';
 
 class CartModel extends BaseModel<Map<int, Product>> {
