+++ white_tiger_shop/lib/category/model/categories_model.dart
@@ -1,6 +1,6 @@
 import 'package:white_tiger_shop/category/controller/categories_api.dart';
 import 'package:white_tiger_shop/category/model/entity/category.dart';
-import 'package:white_tiger_shop/common/model/base_model.dart';
+import 'package:white_tiger_shop/core/model/base_model.dart';
 
 class CategoriesModel extends BaseModel<List<Category>?> {
   List<Category>? _categories;
