+++ white_tiger_shop/lib/core/controller/entity/base_response.dart
@@ -1,5 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:white_tiger_shop/common/controller/entity/meta_of_resp.dart';
+import 'package:white_tiger_shop/core/controller/entity/meta_of_resp.dart';
 
 part 'base_response.g.dart';
