+++ white_tiger_shop/lib/product/model/detailed_product_model.dart
@@ -1,4 +1,4 @@
-import 'package:white_tiger_shop/common/model/base_model.dart';
+import 'package:white_tiger_shop/core/model/base_model.dart';
 import 'package:white_tiger_shop/product/controller/products_api.dart';
 import 'package:white_tiger_shop/product/model/entity/product.dart';
 
@@ -13,6 +13,6 @@ class DetailedProductModel extends BaseModel<Product> {
 
   @override
   Future<void> fetch() async {
-    _product = await api.getDetailedProduct(productId!);
+    _product = await api.getDetailedProduct(productId);
   }
 }
