+++ white_tiger_shop/lib/product/model/products_model.dart
@@ -1,4 +1,4 @@
-import 'package:white_tiger_shop/common/model/base_model.dart';
+import 'package:white_tiger_shop/core/model/base_model.dart';
 import 'package:white_tiger_shop/product/controller/products_api.dart';
 import 'package:white_tiger_shop/product/model/entity/product.dart';
 import 'package:white_tiger_shop/product/model/entity/sort_option.dart';
