+++ white_tiger_shop/lib/product/products_list_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:white_tiger_shop/category/model/entity/category.dart';
-import 'package:white_tiger_shop/common/view/base_page.dart';
+import 'package:white_tiger_shop/core/page/base_page.dart';
 import 'package:white_tiger_shop/product/model/data/sort_options.dart';
 import 'package:white_tiger_shop/product/model/products_model.dart';
 import 'package:white_tiger_shop/product/detailed_product_page.dart';
