+++ white_tiger_shop/lib/product/view/products_list_item.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:white_tiger_shop/common/view/networked_image.dart';
+import 'package:white_tiger_shop/core/view/networked_image.dart';
 import 'package:white_tiger_shop/product/model/entity/product.dart';
 
 class ProductsItemView extends StatelessWidget {
